New Design Database Fix code suported new database

+++ database/migrations/2018_11_17_105639_create_guesthouses_table.php
@@ -15,7 +15,7 @@ class CreateGuesthousesTable extends Migration
         Schema::create('guesthouses', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name',50);
-            $table->tinyInteger('zone_id');
+            $table->integer('zone_id');
 
            
         });
